requirements_production.txt
Django==4.2.7
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
openpyxl==3.1.2
